Creación de modelos de respuesta API y mappers

+++ lib/domain/entities/movie.dart
@@ -3,10 +3,10 @@ class Movie{
   final String name;
   final int runtimeInMinutes;
   final int budgetInMillions;
-  final int boxOfficeRevenueInMillions;
+  final double boxOfficeRevenueInMillions;
   final int academyAwardNominations;
   final int academyAwardWins;
-  final int rottenTomatoesScore;
+  final double rottenTomatoesScore;
 
   Movie({required this.id, required this.name, required this.runtimeInMinutes, required this.budgetInMillions, required this.boxOfficeRevenueInMillions, required this.academyAwardNominations, required this.academyAwardWins, required this.rottenTomatoesScore});
 }
